requirements.txt
opencv-python >= 4.5.5.64
tqdm >= 4.63.1
joblib >= 1.1.0
pillow >= 9.1.0

